refactor: use Metadata API for noindex instead of client-side useEffect
