+++ app/(dashboard)/einstellungen/layout.tsx
@@ -1,55 +1,13 @@
-"use client"
+import type { Metadata } from "next"
+import { SettingsLayoutClient } from "./settings-layout-client"
 
-import { useMemo, useEffect } from "react"
-import { SettingsSidebar } from "@/components/settings/sidebar"
-import type { Tab } from "@/types/settings"
-import {
-  User as UserIcon,
-  Lock,
-  CreditCard,
-  DownloadCloud,
-  Info,
-  Monitor,
-  FlaskConical,
-  Mail,
-  Brain,
-} from "lucide-react"
-import { useFeatureFlagEnabled } from "posthog-js/react"
-import { POSTHOG_FEATURE_FLAGS } from "@/lib/constants"
+export const metadata: Metadata = {
+  robots: {
+    index: false,
+    follow: false,
+  },
+}
 
 export default function EinstellungenLayout({ children }: { children: React.ReactNode }) {
-  useEffect(() => {
-    const meta = document.createElement('meta')
-    meta.name = 'robots'
-    meta.content = 'noindex, nofollow'
-    document.head.appendChild(meta)
-    return () => meta.remove()
-  }, [])
-
-  const mailsEnabled = useFeatureFlagEnabled(POSTHOG_FEATURE_FLAGS.MAILS_TAB)
-  const aiAgentEnabled = useFeatureFlagEnabled(POSTHOG_FEATURE_FLAGS.MIETEVO_AI_AGENT)
-
-  const tabs = useMemo<Omit<Tab, "content">[]>(
-    () => [
-      { value: "profil", label: "Profil", icon: UserIcon },
-      { value: "sicherheit", label: "Sicherheit", icon: Lock },
-      { value: "abo", label: "Abo", icon: CreditCard },
-      ...(mailsEnabled ? [{ value: "mail", label: "E-Mail", icon: Mail }] : []),
-      { value: "darstellung", label: "Darstellung", icon: Monitor },
-      { value: "datenexport", label: "Datenexport", icon: DownloadCloud },
-      { value: "vorschau", label: "Vorschau", icon: FlaskConical },
-      ...(aiAgentEnabled ? [{ value: "ki", label: "KI", icon: Brain }] : []),
-      { value: "mietevo", label: "Mietevo", icon: Info },
-    ],
-    [mailsEnabled, aiAgentEnabled],
-  )
-
-  return (
-    <div className="flex h-full gap-3 p-4">
-      <SettingsSidebar tabs={tabs} />
-      <div className="flex-1 overflow-y-auto min-w-0 p-4">
-        {children}
-      </div>
-    </div>
-  )
+  return <SettingsLayoutClient>{children}</SettingsLayoutClient>
 }
